requirements.txt
Django>=5.2,<6
djangorestframework>=3.15
django-cors-headers>=4.6
psycopg[binary]>=3.2
django-ipware>=7.0
user-agents>=2.2
requests>=2.32
